backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15
django-cors-headers>=4.3
django-filter>=24.0
djangorestframework-simplejwt>=5.3
psycopg[binary]>=3.1
celery[redis]>=5.3
django-celery-results>=2.5
scikit-learn>=1.4
xgboost>=2.0
pandas>=2.2
numpy>=1.26
joblib>=1.3
anthropic>=0.40
gunicorn>=21.2
pytest>=8.0
pytest-django>=4.8
python-dotenv>=1.0
